backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
django-cors-headers==4.1.0
django-environ==0.11.2
djangorestframework==3.15.1
environ==1.0
PyMySQL==1.1.1
pytz==2025.2
sqlparse==0.4.4
typing-extensions==4.7.1